Added if statements to advise user on chems to add to pool

+++ PoolSolutions/Models/Chem.cs
@@ -8,7 +8,7 @@
         }
         public int PoolID { get; set; }
         public double PH { get; set; }
-        public int Chlorine { get; set; }
+        public double Chlorine { get; set; }
         public int Bromine { get; set; }
         public int Alkalinity { get; set; }
         public int Stabilizer { get; set; }
